Correct various types warnings
- Move struct fiber_args in tntsched.h
- Replace some unsigned int with socklen_t
- ...

+++ include/frame.h
@@ -19,7 +19,7 @@
 struct vector_frame;
 
 struct frame {
-    unsigned short size;
+    size_t size;
     void *frame;
     void *raw_packet;
 };
